refactor(web): migrate storybook component
